First commit of some simple unit tests of two of the new methods in RemoteCurrencyServer, currencies and exchangeRates. Commit also includes changes to project, method declarations in headers files, and ordering of header imports, in order to build RemoteCurrencyServer and its dependencies in the test project.
At this point, I’m unsure that I’m heading in the right direction with unit testing. The two simple methods that appear in the tests have external dependencies, and it’s only the dependencies that are being tested. Which seems like it must be violating some  principle of unit testing. The remoteCurrencyServerConnection:currenciesCommand: requires a client/server connection, which is a much larger dependency. An obvious approach for testing this particular method would be to factor out the logic to be tested, from its external dependencies. However, in this case, that leaves no logic to factor out. I would say this method is not a good candidate for unit testing.